Replace paused with entity_status
https://twittercommunity.com/t/ads-api-version-2/90360

says:

" The new entity_status parameter (already included as a response attribute)
will be writable on campaigns and line items. This controls whether a
campaign or line item is active, in draft mode, or is paused.

The paused parameter is no longer accepted nor returned in the response.
To pause campaigns or line items, use the entity_status parameter. "